Jason ingham branch (#16)
* Add some custom annotations

This gives you access to several annotations that you’ll need for the
assignment, as well as adding a few tests. You’ll want these tests to pass
when you’re done with the assignment

* An API between the data storage system and the compute engine

The data
storage system is responsible for reading from and writing to user-specified input
and output sources. Added @ProcessAPI and @ProcessAPIPrototype to the
interface for the API and the prototype method